VESTRO-BACKEND/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1

# Scrapers
playwright==1.44.0
httpx==0.27.0
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.1

# ML
scikit-learn==1.4.2
shap==0.45.1
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

# Scheduling
apscheduler==3.10.4

# Utils
python-dotenv==1.0.1
tenacity==8.3.0
psycopg2-binary
asyncpg